docs: Refactors + DocFX auto-docs system
This commit rolls in several changes:

 - Name refactorings for the high-level OPA clients.
 - Small bugfix in OpaClient.
 - Build system is now fully owned by us, not Speakeasy.
 - Docs builds are now fully owned by us.
 - Docs are now powered by DocFX.
 - Docs have a CI job that triggers on merges to `main`.
