--- Styra/OpenApi/Styra.OpenApi.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <IsPackable>true</IsPackable>
-    <PackageId>Styra.OpenApi</PackageId>
-    <Version>0.6.0</Version>
-    <Authors>Styra</Authors>
-    <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <PackageReadmeFile>README.md</PackageReadmeFile>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <NoWarn>1591</NoWarn>
-    <RepositoryUrl>https://github.com/StyraInc/opa-csharp</RepositoryUrl>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Include="../README.md" Pack="true" PackagePath="/"/>
-    <None Include="../docs/*" Pack="true" PackagePath="/docs"/>
-
-    <PackageReference Include="newtonsoft.json" Version="13.0.3" />
-    <PackageReference Include="nodatime" Version="3.1.9" />
-  </ItemGroup>
-</Project>
