+++ Styra/Styra.csproj
@@ -13,8 +13,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <None Include="../README.md" Pack="true" PackagePath="/"/>
-    <None Include="../docs/*" Pack="true" PackagePath="/docs"/>
+    <None Include="../README.md" Pack="true" PackagePath="/" />
+    <None Include="../docs/*" Pack="true" PackagePath="/docs" />
 
     <PackageReference Include="newtonsoft.json" Version="13.0.3" />
     <PackageReference Include="nodatime" Version="3.1.9" />
