+++ docs/.gitignore
@@ -1,0 +1,2 @@
+_site/*
+api/*
