--- docs/Models/Components/Explain.md
@@ -1,11 +1,0 @@
-# Explain
-
-
-## Values
-
-| Name    | Value   |
-| ------- | ------- |
-| `Notes` | notes   |
-| `Fails` | fails   |
-| `Full`  | full    |
-| `Debug` | debug   |
