--- docs/Models/Components/GzipAcceptEncoding.md
@@ -1,8 +1,0 @@
-# GzipAcceptEncoding
-
-
-## Values
-
-| Name   | Value  |
-| ------ | ------ |
-| `Gzip` | gzip   |
