--- docs/Models/Components/GzipContentEncoding.md
@@ -1,8 +1,0 @@
-# GzipContentEncoding
-
-
-## Values
-
-| Name   | Value  |
-| ------ | ------ |
-| `Gzip` | gzip   |
