--- docs/Models/Components/HealthyServer.md
@@ -1,7 +1,0 @@
-# HealthyServer
-
-
-## Fields
-
-| Field       | Type        | Required    | Description |
-| ----------- | ----------- | ----------- | ----------- |
