--- docs/Models/Components/Input.md
@@ -1,4 +1,0 @@
-# Input
-
-Arbitrary JSON used within your policies by accessing `input`
-
