--- docs/Models/Components/Result.md
@@ -1,4 +1,0 @@
-# Result
-
-The base or virtual document referred to by the URL path. If the path is undefined, this key will be omitted.
-
