--- docs/Models/Components/Revision.md
@@ -1,8 +1,0 @@
-# Revision
-
-
-## Fields
-
-| Field              | Type               | Required           | Description        |
-| ------------------ | ------------------ | ------------------ | ------------------ |
-| `Revision`         | *string*           | :heavy_check_mark: | N/A                |
