--- docs/Models/Components/SuccessfulPolicyEvaluation.md
@@ -1,11 +1,0 @@
-# SuccessfulPolicyEvaluation
-
-
-## Fields
-
-| Field                                                                                                                                                                                                                                      | Type                                                                                                                                                                                                                                       | Required                                                                                                                                                                                                                                   | Description                                                                                                                                                                                                                                |
-| ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
-| `Result`                                                                                                                                                                                                                                   | [Result](../../Models/Components/Result.md)                                                                                                                                                                                                | :heavy_minus_sign:                                                                                                                                                                                                                         | The base or virtual document referred to by the URL path. If the path is undefined, this key will be omitted.                                                                                                                              |
-| `Metrics`                                                                                                                                                                                                                                  | Dictionary<String, *object*>                                                                                                                                                                                                               | :heavy_minus_sign:                                                                                                                                                                                                                         | If query metrics are enabled, this field contains query performance metrics collected during the parse, compile, and evaluation steps.                                                                                                     |
-| `DecisionId`                                                                                                                                                                                                                               | *string*                                                                                                                                                                                                                                   | :heavy_minus_sign:                                                                                                                                                                                                                         | If decision logging is enabled, this field contains a string that uniquely identifies the decision. The identifier will be included in the decision log event for this decision. Callers can use the identifier for correlation purposes.  |
-| `Provenance`                                                                                                                                                                                                                               | [Provenance](../../Models/Components/Provenance.md)                                                                                                                                                                                        | :heavy_minus_sign:                                                                                                                                                                                                                         | Provenance information can be requested on individual API calls and are returned inline with the API response. To obtain provenance information on an API call, specify the `provenance=true` query parameter when executing the API call. |
