--- docs/Models/Errors/ClientError.md
@@ -1,12 +1,0 @@
-# ClientError
-
-Bad Request
-
-
-## Fields
-
-| Field                                         | Type                                          | Required                                      | Description                                   |
-| --------------------------------------------- | --------------------------------------------- | --------------------------------------------- | --------------------------------------------- |
-| `Code`                                        | *string*                                      | :heavy_check_mark:                            | N/A                                           |
-| `Message`                                     | *string*                                      | :heavy_check_mark:                            | N/A                                           |
-| `Errors`                                      | List<[Errors](../../Models/Errors/Errors.md)> | :heavy_minus_sign:                            | N/A                                           |
