--- docs/Models/Errors/Errors.md
@@ -1,10 +1,0 @@
-# Errors
-
-
-## Fields
-
-| Field                                       | Type                                        | Required                                    | Description                                 |
-| ------------------------------------------- | ------------------------------------------- | ------------------------------------------- | ------------------------------------------- |
-| `Code`                                      | *string*                                    | :heavy_check_mark:                          | N/A                                         |
-| `Message`                                   | *string*                                    | :heavy_check_mark:                          | N/A                                         |
-| `Location`                                  | [Location](../../Models/Errors/Location.md) | :heavy_minus_sign:                          | N/A                                         |
