--- docs/Models/Errors/Location.md
@@ -1,10 +1,0 @@
-# Location
-
-
-## Fields
-
-| Field              | Type               | Required           | Description        |
-| ------------------ | ------------------ | ------------------ | ------------------ |
-| `File`             | *string*           | :heavy_check_mark: | N/A                |
-| `Row`              | *long*             | :heavy_check_mark: | N/A                |
-| `Col`              | *long*             | :heavy_check_mark: | N/A                |
