--- docs/Models/Errors/ServerError.md
@@ -1,12 +1,0 @@
-# ServerError
-
-Server Error
-
-
-## Fields
-
-| Field                                                               | Type                                                                | Required                                                            | Description                                                         |
-| ------------------------------------------------------------------- | ------------------------------------------------------------------- | ------------------------------------------------------------------- | ------------------------------------------------------------------- |
-| `Code`                                                              | *string*                                                            | :heavy_check_mark:                                                  | N/A                                                                 |
-| `Message`                                                           | *string*                                                            | :heavy_check_mark:                                                  | N/A                                                                 |
-| `Errors`                                                            | List<[ServerErrorErrors](../../Models/Errors/ServerErrorErrors.md)> | :heavy_minus_sign:                                                  | N/A                                                                 |
