--- docs/Models/Errors/ServerErrorErrors.md
@@ -1,10 +1,0 @@
-# ServerErrorErrors
-
-
-## Fields
-
-| Field                                                             | Type                                                              | Required                                                          | Description                                                       |
-| ----------------------------------------------------------------- | ----------------------------------------------------------------- | ----------------------------------------------------------------- | ----------------------------------------------------------------- |
-| `Code`                                                            | *string*                                                          | :heavy_check_mark:                                                | N/A                                                               |
-| `Message`                                                         | *string*                                                          | :heavy_check_mark:                                                | N/A                                                               |
-| `Location`                                                        | [ServerErrorLocation](../../Models/Errors/ServerErrorLocation.md) | :heavy_minus_sign:                                                | N/A                                                               |
