--- docs/Models/Errors/ServerErrorLocation.md
@@ -1,10 +1,0 @@
-# ServerErrorLocation
-
-
-## Fields
-
-| Field              | Type               | Required           | Description        |
-| ------------------ | ------------------ | ------------------ | ------------------ |
-| `File`             | *string*           | :heavy_check_mark: | N/A                |
-| `Row`              | *long*             | :heavy_check_mark: | N/A                |
-| `Col`              | *long*             | :heavy_check_mark: | N/A                |
