--- docs/Models/Errors/UnhealthyServer.md
@@ -1,8 +1,0 @@
-# UnhealthyServer
-
-
-## Fields
-
-| Field              | Type               | Required           | Description        |
-| ------------------ | ------------------ | ------------------ | ------------------ |
-| `Code`             | *string*           | :heavy_minus_sign: | N/A                |
