--- docs/Models/Requests/ExecutePolicyWithInputRequestBody.md
@@ -1,10 +1,0 @@
-# ExecutePolicyWithInputRequestBody
-
-The input document
-
-
-## Fields
-
-| Field                                                         | Type                                                          | Required                                                      | Description                                                   |
-| ------------------------------------------------------------- | ------------------------------------------------------------- | ------------------------------------------------------------- | ------------------------------------------------------------- |
-| `Input`                                                       | [Input](../../Models/Components/Input.md)                     | :heavy_check_mark:                                            | Arbitrary JSON used within your policies by accessing `input` |
