--- docs/Models/Requests/ExecutePolicyWithInputResponse.md
@@ -1,12 +1,0 @@
-# ExecutePolicyWithInputResponse
-
-
-## Fields
-
-| Field                                                                                                                                             | Type                                                                                                                                              | Required                                                                                                                                          | Description                                                                                                                                       |
-| ------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `ContentType`                                                                                                                                     | *string*                                                                                                                                          | :heavy_check_mark:                                                                                                                                | HTTP response content type for this operation                                                                                                     |
-| `StatusCode`                                                                                                                                      | *int*                                                                                                                                             | :heavy_check_mark:                                                                                                                                | HTTP response status code for this operation                                                                                                      |
-| `RawResponse`                                                                                                                                     | [HttpResponseMessage](https://learn.microsoft.com/en-us/dotnet/api/system.net.http.httpresponsemessage?view=net-5.0)                              | :heavy_check_mark:                                                                                                                                | Raw HTTP response; suitable for custom response parsing                                                                                           |
-| `SuccessfulPolicyEvaluation`                                                                                                                      | [SuccessfulPolicyEvaluation](../../Models/Components/SuccessfulPolicyEvaluation.md)                                                               | :heavy_minus_sign:                                                                                                                                | Success.<br/>The server also returns 200 if the path refers to an undefined document. In this case, the response will not contain a result property.<br/> |
-| `Headers`                                                                                                                                         | Dictionary<String, List<*string*>>                                                                                                                | :heavy_check_mark:                                                                                                                                | N/A                                                                                                                                               |
