--- docs/Models/Requests/HealthResponse.md
@@ -1,11 +1,0 @@
-# HealthResponse
-
-
-## Fields
-
-| Field                                                                                                                                                                               | Type                                                                                                                                                                                | Required                                                                                                                                                                            | Description                                                                                                                                                                         |
-| ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `ContentType`                                                                                                                                                                       | *string*                                                                                                                                                                            | :heavy_check_mark:                                                                                                                                                                  | HTTP response content type for this operation                                                                                                                                       |
-| `StatusCode`                                                                                                                                                                        | *int*                                                                                                                                                                               | :heavy_check_mark:                                                                                                                                                                  | HTTP response status code for this operation                                                                                                                                        |
-| `RawResponse`                                                                                                                                                                       | [HttpResponseMessage](https://learn.microsoft.com/en-us/dotnet/api/system.net.http.httpresponsemessage?view=net-5.0)                                                                | :heavy_check_mark:                                                                                                                                                                  | Raw HTTP response; suitable for custom response parsing                                                                                                                             |
-| `HealthyServer`                                                                                                                                                                     | [HealthyServer](../../Models/Components/HealthyServer.md)                                                                                                                           | :heavy_minus_sign:                                                                                                                                                                  | OPA service is healthy. If the bundles option is specified then all configured bundles have been activated. If the plugins option is specified then all plugins are in an OK state. |
