+++ docs/docfx.json
@@ -1,0 +1,51 @@
+{
+  "build": {
+    "content": [
+      {
+        "exclude": [
+          "_site/**"
+        ],
+        "files": [
+          "**/*.{md,yml}"
+        ]
+      },
+      {
+        "files": [
+          "README.md"
+        ],
+        "src": "../"
+      }
+    ],
+    "globalMetadata": {
+      "_appName": "Styra C# SDK",
+      "_appTitle": "Styra C# SDK",
+      "_enableSearch": true,
+      "pdf": false
+    },
+    "output": "_site",
+    "resource": [
+      {
+        "files": [
+          "images/**"
+        ]
+      }
+    ],
+    "template": [
+      "default",
+      "modern"
+    ]
+  },
+  "metadata": [
+    {
+      "dest": "api",
+      "src": [
+        {
+          "files": [
+            "*.csproj"
+          ],
+          "src": "../Styra"
+        }
+      ]
+    }
+  ]
+}
