+++ docs/index.md
@@ -1,0 +1,5 @@
+---
+_layout: landing
+---
+
+[!INCLUDE [](../README.md)]
