+++ docs/toc.yml
@@ -1,0 +1,2 @@
+- name: API Reference
+  href: api/
