+++ test/SmokeTest.Tests/SmokeTest.Tests.csproj
@@ -26,7 +26,7 @@
 
   <!-- OPA API Reference. -->
   <ItemGroup>
-    <ProjectReference Include="..\..\Styra\OpenApi\Styra.OpenApi.csproj" />
+    <ProjectReference Include="..\..\Styra\Styra.csproj" />
   </ItemGroup>
 
   <!-- Ensure our data files are copied next to the test executables. -->
