+++ test/test.sln
@@ -3,20 +3,29 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SmokeTest.Tests", "SmokeTest.Tests\SmokeTest.Tests.csproj", "{32BE0D80-A8E6-49A3-822B-C73768A1423D}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SmokeTest.Tests", "SmokeTest.Tests\SmokeTest.Tests.csproj", "{0C928A14-22D3-4A60-AA8D-36509236F2A8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{32BE0D80-A8E6-49A3-822B-C73768A1423D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{32BE0D80-A8E6-49A3-822B-C73768A1423D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{32BE0D80-A8E6-49A3-822B-C73768A1423D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{32BE0D80-A8E6-49A3-822B-C73768A1423D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0C928A14-22D3-4A60-AA8D-36509236F2A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0C928A14-22D3-4A60-AA8D-36509236F2A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0C928A14-22D3-4A60-AA8D-36509236F2A8}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{0C928A14-22D3-4A60-AA8D-36509236F2A8}.Debug|x64.Build.0 = Debug|Any CPU
+		{0C928A14-22D3-4A60-AA8D-36509236F2A8}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0C928A14-22D3-4A60-AA8D-36509236F2A8}.Debug|x86.Build.0 = Debug|Any CPU
+		{0C928A14-22D3-4A60-AA8D-36509236F2A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0C928A14-22D3-4A60-AA8D-36509236F2A8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0C928A14-22D3-4A60-AA8D-36509236F2A8}.Release|x64.ActiveCfg = Release|Any CPU
+		{0C928A14-22D3-4A60-AA8D-36509236F2A8}.Release|x64.Build.0 = Release|Any CPU
+		{0C928A14-22D3-4A60-AA8D-36509236F2A8}.Release|x86.ActiveCfg = Release|Any CPU
+		{0C928A14-22D3-4A60-AA8D-36509236F2A8}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
